worker::durable

Struct State

pub struct State { /* private fields */ }

Passed from the runtime to provide access to the Durable Object’s storage as well as various metadata about the Object.

Implementations

impl State

pub fn id(&self) -> ObjectId<'_>

The ID of this Durable Object which can be converted into a hex string using its to_string() method.

pub fn storage(&self) -> Storage

Contains methods for accessing persistent storage via the transactional storage API. See Transactional Storage API for a detailed reference.

pub fn container(&self) -> Option<Container>

pub fn wait_until<F>(&self, future: F)
where F: Future<Output = ()> + 'static,

pub fn _inner(self) -> DurableObjectState

pub fn accept_web_socket(&self, ws: &WebSocket)

pub fn accept_websocket_with_tags(&self, ws: &WebSocket, tags: &[&str])

pub fn get_websockets(&self) -> Vec<WebSocket>

pub fn get_websockets_with_tag(&self, tag: &str) -> Vec<WebSocket>

pub fn get_tags(&self, websocket: &WebSocket) -> Vec<String>

Retrieve tags from a hibernatable websocket

pub fn set_websocket_auto_response(&self, pair: &WebSocketRequestResponsePair)

pub fn get_websocket_auto_response( &self, ) -> Option<WebSocketRequestResponsePair>