worker

Struct AbortSignal

pub struct AbortSignal { /* private fields */ }

An interface representing a signal that can be passed to cancellable operations, primarily a Fetch request.

Implementations

impl AbortSignal

pub fn aborted(&self) -> bool

A bool indicating if the operation that the signal is used for has been aborted.

pub fn reason(&self) -> Option<JsValue>

The reason why the signal was aborted.

pub fn abort() -> Self

Creates a AbortSignal that is already aborted.

pub fn abort_with_reason(reason: impl Into<JsValue>) -> Self

Creates a AbortSignal that is already aborted with the provided reason.