worker

Struct Object

pub struct Object { /* private fields */ }

Object is created when you put an object into a Bucket. Object represents the metadata of an object based on the information provided by the uploader. Every object that you put into a Bucket will have an Object created.

Implementations

impl Object

pub fn key(&self) -> String

pub fn version(&self) -> String

pub fn size(&self) -> u64

pub fn etag(&self) -> String

pub fn http_etag(&self) -> String

pub fn uploaded(&self) -> Date

pub fn http_metadata(&self) -> HttpMetadata

pub fn checksum(&self) -> R2Checksums

pub fn custom_metadata(&self) -> Result<HashMap<String, String>>

pub fn range(&self) -> Result<Range>

pub fn body(&self) -> Option<ObjectBody<'_>>

pub fn body_used(&self) -> Option<bool>

pub fn write_http_metadata(&self, headers: Headers) -> Result<()>