worker

Trait FromRequest

pub trait FromRequest: Sized {
    // Required method
    fn from_raw(request: Request) -> Result<Self, impl Into<Box<dyn Error>>>;
}

A trait used to represent any viable Request type that can be used in the Worker. The only requirement is that it be convertible from a web_sys::Request.

Required Methods

fn from_raw(request: Request) -> Result<Self, impl Into<Box<dyn Error>>>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.