worker

Trait IntoResponse

pub trait IntoResponse {
    // Required method
    fn into_raw(self) -> Result<Response, impl Into<Box<dyn Error>>>;
}

A trait used to represent any viable Response type that can be used in the Worker. The only requirement is that it be convertible to a web_sys::Response.

Required Methods

fn into_raw(self) -> Result<Response, impl Into<Box<dyn Error>>>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.